+++ src/core/_01_rss_ingest.py
@@ -3,7 +3,6 @@ import logging
 import os
 
 # third-party imports
-from dotenv import load_dotenv
 import feedparser
 import polars as pl
 from feedparser import FeedParserDict
@@ -16,27 +15,8 @@ from src.data_models import MediaDataFrame
 # load environment variables
 # ------------------------------------------------------------------------------
 
-# get reel-driver env vars
-load_dotenv(override=True)
-
-log_level = os.getenv('LOG_LEVEL', default="INFO")
-
-if log_level == "INFO":
-    logging.basicConfig(
-        level=logging.INFO,
-        format='%(asctime)s - %(levelname)s - %(message)s',
-        datefmt='%Y-%m-%d %H:%M:%S'
-    )
-    logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
-    logging.getLogger("paramiko").setLevel(logging.WARNING)
-elif log_level == "DEBUG":
-    logging.basicConfig(
-        level=logging.DEBUG,
-        format='%(asctime)s - %(levelname)s - %(module)s - %(funcName)s - %(lineno)d - %(message)s',
-        datefmt='%Y-%m-%d %H:%M:%S'
-    )
-    logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
-    logging.getLogger("paramiko").setLevel(logging.INFO)
+# log config
+utils.setup_logging()
 
 # ------------------------------------------------------------------------------
 # rss ingest helper functions